refactor: adjusting packages
