+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/controller/TodoController.java
@@ -3,9 +3,9 @@ package br.com.jpcchaves.infrastructure.controller;
 import br.com.jpcchaves.core.domain.PaginatedResponse;
 import br.com.jpcchaves.core.domain.PaginationRequest;
 import br.com.jpcchaves.infrastructure.controller.resources.ITodoResource;
-import br.com.jpcchaves.infrastructure.dto.TodoRequestDTO;
-import br.com.jpcchaves.infrastructure.dto.TodoResponseDTO;
-import br.com.jpcchaves.infrastructure.dto.UpdateTodoStatusDTO;
+import br.com.jpcchaves.infrastructure.dto.todo.TodoRequestDTO;
+import br.com.jpcchaves.infrastructure.dto.todo.TodoResponseDTO;
+import br.com.jpcchaves.infrastructure.dto.todo.UpdateTodoStatusDTO;
 import br.com.jpcchaves.infrastructure.service.todo.TodoService;
 import jakarta.validation.Valid;
 import jakarta.websocket.server.PathParam;
