+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/dto/category/CategoryRequestDTO.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.infrastructure.dto;
+package br.com.jpcchaves.infrastructure.dto.category;
 
 public class CategoryRequestDTO {
   private String name;
