+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/mapper/TodoResponseDTOMapper.java
@@ -1,7 +1,7 @@
 package br.com.jpcchaves.infrastructure.mapper;
 
 import br.com.jpcchaves.core.domain.Todo;
-import br.com.jpcchaves.infrastructure.dto.TodoResponseDTO;
+import br.com.jpcchaves.infrastructure.dto.todo.TodoResponseDTO;
 import org.modelmapper.PropertyMap;
 
 public class TodoResponseDTOMapper extends PropertyMap<Todo, TodoResponseDTO> {
