+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/service/category/CategoryService.java
@@ -1,7 +1,7 @@
 package br.com.jpcchaves.infrastructure.service.category;
 
-import br.com.jpcchaves.infrastructure.dto.CategoryRequestDTO;
-import br.com.jpcchaves.infrastructure.dto.CategoryResponseDTO;
+import br.com.jpcchaves.infrastructure.dto.category.CategoryRequestDTO;
+import br.com.jpcchaves.infrastructure.dto.category.CategoryResponseDTO;
 import java.util.List;
 
 public interface CategoryService {
