+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/service/todo/TodoService.java
@@ -3,8 +3,8 @@ package br.com.jpcchaves.infrastructure.service.todo;
 import br.com.jpcchaves.core.domain.PaginatedResponse;
 import br.com.jpcchaves.core.domain.PaginationRequest;
 import br.com.jpcchaves.core.domain.enums.TodoStatus;
-import br.com.jpcchaves.infrastructure.dto.TodoRequestDTO;
-import br.com.jpcchaves.infrastructure.dto.TodoResponseDTO;
+import br.com.jpcchaves.infrastructure.dto.todo.TodoRequestDTO;
+import br.com.jpcchaves.infrastructure.dto.todo.TodoResponseDTO;
 import java.util.List;
 
 public interface TodoService {
